Now uses wails options.ErrorFormatter as an error converter from backend to frontend. Created new terrors package so that it can be imported in sub packages.
